requirements.txt
streamlit
pandas
PyPDF2
Pillow
pytesseract
langchain_google_genai
clickhouse_connect
google-generativeai
